+++ src/visiomode/stimuli/__init__.py
@@ -87,7 +87,7 @@ class Stimulus(pg.sprite.Sprite, mixins.BaseClassMixin, mixins.WebFormMixin):
 
     def get_details(self):
         """Returns a dictionary of stimulus attributes."""
-        return dict()
+        return {"id": self.get_identifier, "common_name": self.get_common_name}
 
     def generate_new_trial(self):
         """Regenerate stimuli for a fresh trial"""