+++ src/visiomode/stimuli/variable_contrast_grating.py
@@ -31,4 +31,8 @@ class VariableContrastGrating(grating.Grating):
         self.area = self.screen.get_rect()
 
     def get_details(self):
-        return {"trial_contrast": self.trial_contrast}
+        return {
+            "id": self.get_identifier,
+            "common_name": self.get_common_name,
+            "trial_contrast": self.trial_contrast,
+        }